switched to kitty @ launch

+++ music/album_cover.sh
@@ -10,10 +10,11 @@ echo "$file_name" >> /tmp/some_stuff.txt
 
 name=$(head -n 1 /tmp/some_stuff.txt)
 
-rm /tmp/some_stuff.txt
+if [[ -f /tmp/some_stuff.txt ]]; then
+    rm /tmp/some_stuff.txt
+fi
 
 name="${name%.*}"
 COVER="/home/oliver/$name.webp"
-
 
 kitty @ send-text --match title:cover 'clear && kitten icat --silent --scale-up --clear --stdin=no --align center "/home/oliver/'$name'.webp" \r'
